Remove `maxFiles` limitation.
It does not make sense that a function limits the number of file
that the entire process can use, and does not undo this limitation
when the function is done.

Other functions that run afterwards may need more files,
and the number may change over time as the code develops.
Those will fail with `EMFILE (too many open files)` without further
explanation.

It also makes it very difficult to develop any addition to PoissonRecon
that runs e.g. as another thread in the same process (e.g. thread
that does progress logging or similar).